+++ CHANGELOG.md
@@ -25,3 +25,8 @@
 ## 1.0.4
 
 - Load file improvements, way to set http-header-fields, needed for playing data from streaming services.
+
+## 1.0.5 (Draft)
+
+- max-volume added (contribution by: https://github.com/byter11)
+- multiple MPV instances supported (set webport and webportrangeend on your settings)